SDK-1325
Specific GetDeviceName for Android and IOS mobile platforms. email_confirmation_required bug fix. Int64 instead of int32 for DeviceId

+++ Source/XsollaLogin/Public/XsollaLoginTypes.h
@@ -445,7 +445,7 @@ struct FXsollaUserDevice
 	FString device;
 
 	UPROPERTY(BlueprintReadWrite, Category = "User Device")
-	int32 id;
+	int64 id;
 
 	UPROPERTY(BlueprintReadWrite, Category = "User Device")
 	FString last_used_at;
